Move all maps from Menu.h to the source files that use them.
